refactor(modals): restructure collection modal components and update imports
